+++ src/main/kotlin/taxonomy/service/RandomTournamentScheduler.kt
@@ -51,7 +51,7 @@ class RandomTournamentScheduler(
                 val isModelA = key.first == ps.modelA
                 val winsFirst = if (isModelA) ps.winsA else ps.winsB
                 stats.sumX = winsFirst + 0.5 * ps.ties
-                stats.n = ps.totalComparisons
+                stats.n = ps.totalComparisons.toInt()
             }
 
             leafArena
